requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
python-dotenv==1.0.1
requests==2.32.3
qrcode==7.4.2
Pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==2.1.1
